+++ src/virtual.d.ts
@@ -1,0 +1,4 @@
+declare module 'virtual:cantare-config' {
+  const config: import('./lib/site').SiteConfig;
+  export default config;
+}